Move getHourlyPrice from ResourceIconts to resourceUtils
Refs #630.
